refactor(tests): split test_models.py into a focused models package (#443)
tests/teatree_core/test_models.py was 1306 LOC and the largest remaining
standalone oversized test module after #896 (management_commands/),
#912 (sync/) and #913 (utils/). It exercised six distinct model
concerns (Ticket, per-phase auto-dispatch, Worktree, Session, Task,
PullRequest/admin) in a single file.

This splits it into a focused tests/teatree_core/models/ package, one
module per concern, with the five ticket-advancement / git-repo-setup
helpers lifted verbatim into a shared _shared.py.

Pure verbatim reorganization: every class and helper is AST-identical
to the original (verified programmatically), only import paths changed
and narrowed per file. Zero production code touched, zero assertions
changed. Same 80 tests collected and passing before and after.

Relates-to #443
